requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
requests
aiogram==3.10.0
pydantic==2.6.4
